back/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
fastapi==0.92.0
frozenlist==1.4.0
h11==0.14.0
httptools==0.6.0
idna==3.4
multidict==6.0.4
openai==0.27.0
pydantic==1.10.12
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.28.2
sniffio==1.3.0
starlette==0.25.0
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2
